+++ nucleotides/metrics.py
@@ -3,6 +3,9 @@ Provides functions for converting the streamed cgroup data produced from monitor
 the Docker container into metrics that may be uploaded to the nucleotides API.
 """
 
+# Make all builtin functions accessible to mapping lifts
+from __builtin__ import *
+
 import funcy, os
 import ruamel.yaml as yaml
 
@@ -11,11 +14,15 @@ from functools import partial
 import biobox.cgroup    as cgroup
 import nucleotides.util as util
 
-SAMPLING_INTERVAL      = 15
-BYTE_TO_MIBIBYTE       = 1.0 / 1024 ** 2
-SECONDS                = 1
-NANOSECONDS_TO_SECONDS = 1e-9
 
+def round_to_3(x):
+    return round(x, 3)
+
+def byte_to_mibibyte(x):
+    return x * (1.0 / 1024 ** 2)
+
+def nanoseconds_to_seconds(x):
+    return x * 1e-9
 
 def time_diff(xs):
     return cgroup.time_diff_in_seconds(xs[0], xs[-1])
@@ -25,17 +32,6 @@ def parse_quast_value(x):
     return quast_mapping[x] if x in quast_mapping else x
 
 
-CGROUP_JMESPATHS = {
-    "total_memory_usage_in_mibibytes"          : [funcy.last, "memory_stats.max_usage", BYTE_TO_MIBIBYTE],
-    "total_rss_in_mibibytes"                   : [max, "memory_stats.stats.total_rss", BYTE_TO_MIBIBYTE],
-    "total_cpu_usage_in_seconds"               : [funcy.last, "cpu_stats.cpu_usage.total_usage", NANOSECONDS_TO_SECONDS],
-    "total_cpu_usage_in_seconds_in_kernelmode" : [funcy.last, "cpu_stats.cpu_usage.usage_in_kernelmode", NANOSECONDS_TO_SECONDS],
-    "total_cpu_usage_in_seconds_in_usermode"   : [funcy.last, "cpu_stats.cpu_usage.usage_in_usermode", NANOSECONDS_TO_SECONDS],
-    "total_read_io_in_mibibytes"               : [funcy.last, "sum(blkio_stats.io_service_bytes_recursive[?op=='Read'].value)", BYTE_TO_MIBIBYTE],
-    "total_write_io_in_mibibytes"              : [funcy.last, "sum(blkio_stats.io_service_bytes_recursive[?op=='Write'].value)", BYTE_TO_MIBIBYTE],
-    "total_wall_clock_time_in_seconds"         : [time_diff,  "read", SECONDS]}
-
-
 def get_expected_keys_from_mapping_file(name):
     """
     Returns the list of metrics that should be collected based on the metric file name
@@ -43,7 +39,6 @@ def get_expected_keys_from_mapping_file(name):
     path = os.path.join('mappings', name + '.yml')
     mappings = yaml.safe_load(util.get_asset_file_contents(path))
     return list(map(lambda x: x['key'], mappings))
-
 
 
 def parse_metrics(app, metrics, mappings):
@@ -102,33 +97,17 @@ def are_metrics_complete(app, expected, collected):
     return not missing_metrics
 
 
-def extract_metric(doc, path):
+def check_90_percent_real_values(x):
     """
-    Given a JSON document, fetches fields out for a given jmespath path, if
-    >= 10% of the values are None, returns None.
+    Given a list of values, if >= 10% of the values are None, returns None.
     """
-    import jmespath
-    f = funcy.rcompose(
-            partial(map, jmespath.compile(path).search),
-            partial(filter, lambda x: x is not None))
+    collection_threshold = 0.15
 
-    values = f(doc)
-    perc_missing = round(1 - float(len(values)) /  len(doc), 2)
-    return None if perc_missing >= 0.1 else values
+    if len(x) == 0:
+        return [0.0]
 
-
-
-def parse_runtime_metrics(metrics):
-    """
-    Given a list of cgroup dictionaries, parses them into a single dictionary of
-    nucleotides metrics that can be uploaded to the nucleotides API. Ignores metrics
-    where more than 10% of the values are missing.
-    """
-    def parse(acc, (name, (f, path, units))):
-        values = extract_metric(metrics, path)
-        if values:
-            acc.append([name, round(f(values) * units, 3)])
-        return acc
-
-    nucleotides_metrics = dict(reduce(parse, CGROUP_JMESPATHS.iteritems(), []))
-    return nucleotides_metrics
+    percent_none_values = len(funcy.remove(funcy.notnone, x)) / len(x)
+    if percent_none_values >= collection_threshold:
+        return [0.0]
+    else:
+        return x