+++ nucleotides/task/short_read_assembler.py
@@ -1,6 +1,8 @@
 import os.path, shutil, funcy
 
+import ruamel.yaml            as yaml
 import nucleotides.metrics    as met
+import nucleotides.util       as util
 import nucleotides.filesystem as fs
 
 from nucleotides.task.task_interface import TaskInterface
@@ -23,11 +25,17 @@ class ShortReadAssemblerTask(TaskInterface):
     def collect_metrics(self, app):
         import json, gzip
         path = fs.get_task_file_path(app, "outputs/container_runtime_metrics/metrics.json.gz")
-        if os.path.isfile(path):
-            with gzip.open(path) as f:
-                return met.parse_runtime_metrics(json.loads(f.read()))
-        else:
+
+        if not os.path.isfile(path):
             return {}
+
+        with gzip.open(path) as f:
+            raw_metrics = json.loads(f.read())
+
+        mapping_file = os.path.join('mappings', self.metric_mapping_file(app) + '.yml')
+        mapping      = yaml.safe_load(util.get_asset_file_contents(mapping_file))
+        return met.parse_metrics(app, raw_metrics, mapping)
+
 
     def successful_event_output_files(self):
         return set(["contig_fasta"])